refactor(direct-control): extract progress dashboard source
Move the progress dashboard embedded App UI source into src/play/progression/progress-dashboard.ts while keeping the public wrapper and command builder in the facade. Keep the move as source ownership only; this does not claim live-game runtime proof.

Record the progression read atom status in the support modularization OpenSpec workstream, including the remaining tactical/read source owners.

Validation: bun run --cwd packages/civ7-direct-control test -- progression-reads.test.ts; bun run --cwd packages/civ7-direct-control test; bun run --cwd packages/civ7-direct-control check; bun run --cwd packages/civ7-direct-control build; bun run openspec -- validate civ7-support-direct-control-modularization --strict; bun run resources:status; git diff --check.
